+++ Areas/Member/Models/CartItem.cs
@@ -7,11 +7,11 @@ namespace PROG3050_HMJJ.Areas.Member.Models
     {
         public int GameId { get; set; }
         public string GameName { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
         public string OrderType { get; set; }  // Added this to indicate the type of order ("Physical" or "Digital")
         public bool IsPurchased { get; set; }
-        public int TotalPrice { get { return Price * Quantity; } }
+        public decimal TotalPrice { get { return Price * Quantity; } }
 
     }
 }
